test(chat): split queued turn regression coverage
Summary:
- move the new background queued-turn regressions into a dedicated chat test
  file so the web file-budget lint stays green

Rationale:
- the workflow-aligned web CI gate enforces a hard max-lines cap for the turns
  test file, so splitting the new cases keeps the regression coverage without
  breaking lint

Tests:
- corepack pnpm --dir web exec vitest run src/lib/features/chat/project-conversation-controller-actions.test.ts src/lib/features/chat/project-conversation-controller-queued-turns.test.ts src/lib/features/chat/project-conversation-controller-tabs.test.ts src/lib/features/chat/project-conversation-controller-turns.test.ts src/lib/features/chat/project-conversation-panel-tabs.test.ts src/lib/features/chat/project-conversation-panel.test.ts src/lib/features/chat/project-conversation-controller-restore.test.ts
- PLAYWRIGHT_WEB_PORT=4273 corepack pnpm --dir web exec playwright test --reporter=dot --quiet
